requirements.txt
# HOI4 파싱
mwparserfromhell==0.6.6

# 위키 API
requests==2.31.0
SPARQLWrapper==2.0.0

# 이미지 처리
Pillow==10.2.0
rembg==2.0.67
opencv-python==4.9.0.80
mediapipe==0.10.21

# 데이터 모델
pydantic==2.11.0

# 웹 UI
streamlit==1.31.0

# 테스트
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# 유틸리티
click>=8.1.8
tqdm==4.66.1
loguru==0.7.2
python-dotenv==1.0.0

# AI / 에이전트
anthropic>=0.40.0
google-generativeai>=0.8.0
google-genai>=0.3.0
ddgs>=7.0.0
tavily-python>=0.5.0
